requirements.txt
common-dibbs
django>=1.11,<2.0
djangorestframework>=3.5,<3.6
pyjwt==1.4.2
